requirement.txt
pandas
numpy
scikit-learn
mlflow
dvc

dagshub
flask
python-dotenv
joblib
pyyaml